Bazel build files for allowing poprithms to build with Tensorflow
Summary:
The Bazel rule just slurps in all of the CPP and H files and treats
them as a single library.  If this becomes insufficient in the future
then we can break the system into multiple independent libraries.

Test Plan: Build locally with TF, including one of the header files in TF.

Reviewers: #poprithms, #popart, jamesn

Reviewed By: #poprithms, #popart, jamesn

Differential Revision: https://phabricator.sourcevertex.net/D21013