Simplify minSumLiveness by removing user provided Markov probabilities
Summary:
It was found in 6 regression suite Graphs that the complex code to randomly choose
between 3 approaches to choosing the number of Ops to shift did not provide any measurable
speed-up to converging to a local minimum.
The code is removed in this diff, although the final deprecation of the API has been postponed.

Test Plan: All tests pass with this simplification

Reviewers: #poprithms, evan

Reviewed By: #poprithms, evan

Subscribers: evan, fabiant

Maniphest Tasks: T16682

Differential Revision: https://phabricator.sourcevertex.net/D22130