Get boost dependencies from boost_scripts repo
Summary:
Having this list here has been causing issues with the release builder so rBOOSTSCRIPTS has been updated to contain the most up-to-date list of dependencies (including graph and context) here: D27897.

Removing this now so that we don't have two lists that can diverge and, if there has to be one place for all repos, it seems better to put it in rBOOSTSCRIPTS than rPOPRITHMS :)

Ref T17854.

Test Plan: ci still builds, @jamiehe manually tested that this fixes the issue seen in the release builder.

Reviewers: jamesn, george, evan, #poprithms

Reviewed By: jamesn, #poprithms

Subscribers: jamiehe

Maniphest Tasks: T17854

Differential Revision: https://phabricator.sourcevertex.net/D27898