Add POPRITHMS_WERROR cmake option
Summary:
Allows poprithms to be configured with -Werror enabled/disabled.

In building (just) the poprithms project, user can configure with
```
/local/build/ cmake . -DPOPRITHMS_WERROR=1
```
to convert all compiler warnings in library code into errors.
Alternatively,
```
/local/build/ cmake . -DPOPRITHMS_WERROR=0
```
will leave all compiler warnings as just warnings.

The default is POPRITHMS_WERROR=1: warnings=errors.

A follow-up task is to expose the option to super-builds, in particular if
a user wishes to toggle the option when building the poplar or popart views.
The follow-up task for this is: T22655.

Test Plan: Option tested locally

Reviewers: #poprithms, anthonyb, samk, george

Reviewed By: #poprithms, anthonyb, george

Maniphest Tasks: T22048

Differential Revision: https://phabricator.sourcevertex.net/D28029