Merge base gbase::Node class and alias::Node classes
Summary:
The base class is complicating design decisions for the second use case,
which is still a moving target. We can return to this at a later
point when it is clearer what the base class should contain. For now, keeping
the design simple by removing the base class.

Reviewers: shirazb, #poprithms, markk

Reviewed By: #poprithms, markk

Subscribers: markk

Maniphest Tasks: T24945

Differential Revision: https://phabricator.sourcevertex.net/D30176