Implement Shape::squeeze for a subset of singleton dimensions.
Summary:
There is currently a method to squeeze out all singleton
dimensions in a Shape. This patch allows only a subset of singleton
dimensions to be squeezed out.

If a non-singleton dimension is in the squeeze set, an error is thrown.
Duplicates are allowed in the squeeze set (which is a std::vector).

Zero dimensions are not squeezed out (would change number of elements of Shape)

Test Plan: New test added

Reviewers: #poprithms, george

Reviewed By: #poprithms, george

Subscribers: evan

Maniphest Tasks: T26826

Differential Revision: https://phabricator.sourcevertex.net/D32303