Implement view changing operations for compute::host::Tensor
Summary:
Given a Shape, and a std::vector with corresponding row-major
data, or a pointer to the first element of a contiguous
row-major buffer, perform transformations such as DimShuffle,
Slice, Concat, and Expand, to create a new std::vector corrsesponding
to the row-major output.

This is a piece of https://phabricator.sourcevertex.net/D32121

Test Plan: Tests for each transformation included.

Reviewers: #poprithms, shirazb, matthewha

Reviewed By: #poprithms, shirazb, matthewha

Subscribers: matthewha, shirazb

Maniphest Tasks: T26981

Differential Revision: https://phabricator.sourcevertex.net/D32484