A util class to "wrap" unique_ptrs of cloneable types.
Summary:
Maintaining bug free constructors and assignment operators for large
classes can be hard work, and the C++ "= default" versions can save
alot of this effort. But if a class has a single non-copyable member, like
a std::unique_ptr, the defaults are not valid.

This patch introduces a template wrapper around std::unique_ptr which makes
it copyable. Wrapping unique_ptrs in this, means that containing classes can
again use the C++ defaults.

The constraints on a class Foo to be wrappable in this class:
- must have a clone() method which returns a std::unique_ptr.
- must have operator==(const Foo &) const implemented

Test Plan: A new test illustrating the use is included in this diff.

Reviewers: george, #poprithms, shirazb, evan, markk

Reviewed By: #poprithms, evan, markk

Subscribers: evan, markk

Maniphest Tasks: T31845

Differential Revision: https://phabricator.sourcevertex.net/D38051