Ability to apply a chain::Chain to a host::Tensor
Summary:
Two different poprithms projects finally in this diff:
1) the host::Tensor class is for "eager" mode computation on a CPU,
   quite similar to a numpy array.
2) the chain::Chain class is a linked list of view-changing operations.
With this diff, you can apply a Chain to a host:Tensor, and get a new
host::Tensor out. Eventually, you will be able to apply a Chain to a
poplar::Tensor. The main motivation for this is testing of unwinding
(coming later).

Test Plan: new test added. Requires D39603 to pass.

Reviewers: markk, #poprithms

Reviewed By: markk, #poprithms

Maniphest Tasks: T33387

Differential Revision: https://phabricator.sourcevertex.net/D39620