Class to summarize times spents in different scopes
Summary:
The class can be thought of as a set of stopwatches, where only 1
stopwatch can be on at a time. It is possible for no stopwatches to be on.
The final summay which this class produces consists of
1) the time in seconds in each scope (time on each stopwatch)
2) the time since the object was created
3) the different (2) - (1) : unaccounted for time.

The idea is that this will make determining which components of a workflow
are slow.

Test Plan: New tests added.

Reviewers: george, #poprithms, markk

Reviewed By: #poprithms, markk

Subscribers: markk

Maniphest Tasks: T34040

Differential Revision: https://phabricator.sourcevertex.net/D40262