RAII design pattern for stopwatches
Summary:
It can be difficult to known when to stop a stopwatch, especially
when there are multiple ways of exiting from the code it is timing.
This task is to use the RAII design so that a stopwatch only needs to
be started - as soon as the scope in which it is started is left, a destructor
stops it. This only works if C++ scopes correspond to the code partitions which
cumulative times, which is normally the case.

Included in this diff
(1) ability to nest stopwatches. Specifically, when a new
stopwatch  starts when there is already another stopwatch on, stop the one
which is on, and then restart it when the new stopwatch stops.
(2) Rename the main class
(3) add threshold on percentage of total time for logging. i.e. don't need
to log the time contribution of components which take up only a small fraction.

Reviewers: markk, #poprithms

Reviewed By: markk, #poprithms

Subscribers: shirazb

Maniphest Tasks: T34276

Differential Revision: https://phabricator.sourcevertex.net/D40449