Constructors for Logger classes which generate random unique ids.
Summary:
The global static object which controls all the Loggers requires that all
Loggers have unique names. The global object is required to make it possible to
set the log level of all Loggers simulateously, and to ensure that their clocks
are in sync (same t=0).

Sometimes a user wants to create a Logger but needs to ensure that no error is
thrown if the name they choose is already taken. This enables allows this option.

The API change backwards compatible, and the default behaviour is still to
throw an error if the user attempts to construct a Logger with an used id.

Reviewers: markk, shirazb, #poprithms

Reviewed By: markk, shirazb, #poprithms

Maniphest Tasks: T34373

Differential Revision: https://phabricator.sourcevertex.net/D40629