Unwinding
Summary:
Standalone implementation of logic for determining good
layouts of input Tensors based on good layouts of internal Tensors.

There is a lengthy description of the problem and the solution
in graph.hpp of the unwinding directory, and in Unwinding.md
and there are many examples of using this class in the unwinding
test directory.

Test Plan: Several new tests added

Reviewers: evan, markk, #poprithms

Reviewed By: markk, #poprithms

Subscribers: matthewha, shirazb, oskarb

Maniphest Tasks: T23020

Differential Revision: https://phabricator.sourcevertex.net/D29615