Ability to permute a DimShuffle and a Reshape in Chain class
Summary:
The Chain
```
Chain({35,2,3}).reshape({5,7,6}).dimShuffle({{2,0,1}})
```
is the same as
```
Chain({35,2,3}).dimShuffle({{1,2,0}}).reshape({6,5,7}).
```
Knowing this, it is possible to canonicalize/simplify the Chain
```
Chain({2,3,35}).dimShuffle({{2,0,1}}}).reshape({5,7,6}).dimShuffle({{2,0,1}})
```
to
```
Chain({2,3,35}).reshape({6,5,7}).
```

This patch implements the swap part of the above, i.e. permuting a Reshape and DimShuffle where possible.

Reviewers: #poprithms, matthewha

Reviewed By: #poprithms, matthewha

Subscribers: matthewha

Maniphest Tasks: T35564

Differential Revision: https://phabricator.sourcevertex.net/D41906