Use depth first search in processWeightSeparated
Summary:
This is a bug fix.

getFilteredSchedule is used incorrectly, as it terminates its search too early.
A simple depth first search is what is required in processWeightSeparated.
An example to illustrate the problem, consider the DAG:

op0 -> op2
op1 -> op2
Starting at op0, getFilteredSchedule will not visit op2, as op2 has the
unsatisfied dependency op1. This is not the behaviour which we want in
processWeightSeparated. A simple depth first search does not
consider the input dependencies of Ops, and so op2 will be visited with
a depth first search.

Test Plan: New test added, which fails before this fix.

Reviewers: #poprithms, shirazb

Reviewed By: #poprithms, shirazb

Subscribers: shirazb

Maniphest Tasks: T37459

Differential Revision: https://phabricator.sourcevertex.net/D43831