A method to determine the (possible) relative position of 1 Op within a subset.
Summary:
Example. Consider the DAG,

```
a --+--> b
    |
    +--> c
```

There are 2 possible schedules, (a, b, c) and (a, c, b).

a appears first in both schedules, and last in neither.
b appears first in neither, and last in 1 of the 2.
c appears first in neither, and last in 1 of the 2.

a, b, and c are thus classified respectively as
a : (IsFirst::Yes, IsFinal::No)
b : (IsFirst::No, IsFinal::Maybe)
c : (IsFirst::No, IsFinal::Maybe)

The method implemented in this diff determines the relative position of any
1 Op in a subset of Ops in the full DAG.

Motivation: There is already a method getRelativePositions(const std::vector<OpId> & subset) which is O(N^2), and processes all OpIds in subset. Profiling the scheduler with Instruments for T36195, this came up as as bottleneck in one stage of the scheduler. At the bottleneck, only 2 of the OpIds in subset actually need to be processed. This method make it possible to process individual OpIds.

Test Plan: New test added.

Reviewers: #poprithms, fabiant

Reviewed By: #poprithms, fabiant

Subscribers: markk

Maniphest Tasks: T37377

Differential Revision: https://phabricator.sourcevertex.net/D43748