Partitioned transitive closures
Summary: The memory used by a TransitiveClosure with N Ops is `2*N*ceil(N/512)*512` bits. This is quadratic in N (with a small constant in the complexity). If the graph was divided by components, the memory footprint would be O(N-ops-in-largest-component^2), with the same constant as before.

Reviewers: #poprithms, markk

Reviewed By: #poprithms, markk

Subscribers: markk

Maniphest Tasks: T40002

Differential Revision: https://phabricator.sourcevertex.net/D46242