Option to avoid opening gates which create multiple input-output aliases
Summary:
This feature can be used to (conservatively) avoid race conditions for inplace
binary elementwise operations. Consider for example:

```
A = input of shape (10,)
out = A + A.reverse();
```

The values of `out` depend on the order in which the elements are added.
Such race conditions never arise if the tensor be adding to (A) does
not alias the the other (A.reverse()).

Test Plan: New tests added.

Reviewers: nejci, fabiant, #poprithms

Reviewed By: nejci, fabiant, #poprithms

Maniphest Tasks: T42084

Differential Revision: https://phabricator.sourcevertex.net/D48040