[schedule::transitiveclosure] Method to bound the duration a set of ops is live
Summary:
This diff implements method which puts bounds on:

```
MAX = max_{schedules} (max-schedule-index(ops) - min-schedule-index(ops)),
```
and

```
MIN = min_{schedules} (max-schedule-index(ops) - min-schedule-index(ops)),
```
for any set of Ops, 'ops'. This will be useful for a future optimization in
the shift scheduler: If an Allocation is always live for the same duration in all
possible schedules, it can safely be removed without changing the optimization problem.
This is exactly the case of `MAX = MIN` when `ops` is all Ops associated to the Allocation.

Also in this diff:
- Some general code quality improvements for the TransitiveClosure class.

Test Plan: new tests added.

Reviewers: shirazb, matthewha, #poprithms

Reviewed By: matthewha, #poprithms

Maniphest Tasks: T43564

Differential Revision: https://phabricator.sourcevertex.net/D49224