[schedule::shift] utility to write scheduling summary reports
Summary:
This diff adds the ability to write json serialized schedule::shift::Graphs,
and a compile time summary, to file. The directory written to
can either be passed to the SummaryWriter constructor, or it can be set
by a user higher up the stack via an environment variable:

`export POPRITHMS_SCHEDULE_SHIFT_WRITE_DIRECTORY=/my/dir`

When the graph is scheduled, 3 files are written, containing respectively
1) the initial (user provided) graph,
2) the final graph (post all the TCOs) which the shifting process acts on,
3) a summary of the time spent in the top-level methods of scheduling.

The motivation for this is to make creating "reproducers" for slow scheduling, easier.

Test Plan: New test added.

Reviewers: #poprithms, george, shirazb

Reviewed By: #poprithms, george

Maniphest Tasks: T43621

Differential Revision: https://phabricator.sourcevertex.net/D49268