[memory::nest] [bug fix] Don't expect the slice of a Region to be 0/1 Regions.
Summary:
Region::slice is a convenience method which calls settSample. A slice is
essentially a settSample, where the slice bounds are prepended onto the setts of
the sliced region. It might therefore seem reasonable to assume that the slice of
a Region is not going to be multiple regions. However, canonicalizations passes
might then convert the potentially complex region into 2 simpler regions. Example:

If the Region is
```
xxx..xxx..xxx..
```

which can be expressed with a depth-1 sett, and it is sliced from 0 to 11:

```
xxx..xxx..xxx..
-----------
```

then the result can either be expressed with 1 region with a depth-2 sett
((11,0,0),(3,2,0)) or with 2 regions with depth-1 setts, (3,2,0) and (1,10,10).

There is no guarantee which will be chosen, so no assumptions should be made.

This essentially means that Region::slice must return DisjointRegions, and not Region.

Test Plan:
- Confirmed that the reproducer in T442367 passes.
- Added unit test which fails before this path, and passes after.

Reviewers: shirazb, markk, #poprithms

Reviewed By: markk, #poprithms

Maniphest Tasks: T44367

Differential Revision: https://phabricator.sourcevertex.net/D49975