Analyze and improve the performance of stopwatches
Summary:
In a test where there are 1e6 switches in total, between 1e2 scopes, with no
work other than switching stopwatches, the time taken before this diff is:

```
Total                   1.307975             n/a       100 %
Unaccounted for         0.893679             n/a        68 %
Accounted for           0.414296             n/a        32 %
```

And then after this diff it is:
```
Total                   0.838901             n/a       100 %
Unaccounted for         0.580622             n/a        69 %
Accounted for           0.258279             n/a        31 %
```

The improvement comes about by replacing a std::string with an integral type.

The 'Unnacounted for' component is mostly generating the summary string from all
the events. So the actual stopwatches switching context only takes about 0.25
seconds to perform 1 million switches.

Reviewers: markk, george, #poprithms

Reviewed By: george, #poprithms

Maniphest Tasks: T44672

Differential Revision: https://phabricator.sourcevertex.net/D50222