[schedule::shift] move caching logic out of constructor
Summary:
The current design, of passing a pointer to a cache to the constructor, is non-intuitive.
Better would be to have a factory function which takes in a cache pointer, and then
either calls into the constructor (in the case of a cache miss) or returns the cached solution.
This is the design in this diff.

Reviewers: shirazb, #poprithms

Reviewed By: shirazb, #poprithms

Maniphest Tasks: T44687

Differential Revision: https://phabricator.sourcevertex.net/D50227