Kahn with priorities in the shift scheduler
Summary:
This diff exposes the option of initializating the schedule with priorities for certain ops.
This functionality already exists in the vanilla scheduler, this diff just exposes it to the
shift scheduler.

The main change is to complement the enum KahnTieBreaker with an optional vector of
priorities.

Test Plan:
We expect that a schedule which is initialized with a more "optimal" schedule should take
fewer rotations to obtain an "optimal" solution. We test this hypothesis in a test.

Reviewers: shirazb, #poprithms

Reviewed By: shirazb, #poprithms

Maniphest Tasks: T46151

Differential Revision: https://phabricator.sourcevertex.net/D51534