[memory::nest] Handle rank-0 region edge case.
Summary:
Regions with rank-0 shape, also called 'scalar' regions, need to be
handled separately. To convince you that this is (unfortunately) true,
consider the following statement:

For any Region r,
    r.totalElms() + r.getComplement().totalElms() =
                                                  = r.shape().nelms()

totalElms is the total number of 'ons' in the region, computed as the
product over dimensions of the number of 'ons' in that dimension. The
base reduction case is totalElms = 1 for rank-0 regions. But if we used
this base case for rank-0 regions, the above statement would not be true
(1 + 1 != 1).

More generally, for any Region whose shape has N elements, we need to be able
to accurately count the number of 'on' elements (0 <= n-ons <= N). For scalars, a
single bit suffices: is the scalar on or off? This diff handles this edge for all of the Region
class methods.

Test Plan: Numerous new tests added

Reviewers: shirazb, #poprithms

Reviewed By: shirazb, #poprithms

Maniphest Tasks: T49687

Differential Revision: https://phabricator.sourcevertex.net/D55062