[memory] Percolate nest::Region::contains up to a inplace::Graph
Summary:
At the base of all alias analysis is the Region class. This class has a method,
'contains', which checks if one Region is the superset of another Region.

This task is percolates this method up the layers of abstraction, all the way to
the inplace::Graph class. Given 2 inplace::Tensors we are now able to say
if one contains the other (in an allocation sense).

Test Plan: New tests added

Reviewers: markk, shirazb, #poprithms

Reviewed By: shirazb, #poprithms

Maniphest Tasks: T50247

Differential Revision: https://phabricator.sourcevertex.net/D55716