[ndarray] createBroadcastOperator view-change helper for multi-dim numpy broadcasts.
Summary:
A helper function to generalize the poplibs 'createBroadcastOperator' function to multiple
dimensions. For createBroadcastOperator, given a tensor t and a **single** non-broadcast
dimension d, a 1-d Tensor of size t.dim(d) is created. However, there might be 2+ dimensions
which are non-broadcast. Consider for example the case where we want to set the layout of a
tensor of shape (1,5,1,7) from a tensor of shape (3,4,5,6,7). To do this, we need to apply some
view-changing dimShuffles and reshapes before we can call createBroadcastOperator.
This diff implements the logic for this in a reusable (and testable) way.

Test Plan: New test added

Reviewers: shirazb, #poprithms

Reviewed By: shirazb, #poprithms

Maniphest Tasks: T51664

Differential Revision: https://phabricator.sourcevertex.net/D57114