[chain] merge sample->fill if sample never removes elements
Summary:
In general, `chain.settSample(r0).settFillInto(r0)` cannot be replaced
by `chain`. However, if `chain` contains Op(s) which ensure that it
the `settSample` never removes elements, then the replacement
(remove sample and fill) is valid. Example where it is not valid:

```
  Chain c({10});
  Region r({10}, {{{{5,5,0}}}});
  c.settSample(r); // this is a slice, like t[0:5];
  c.settFillInto(r);
```
In the above case, applying `c` to a view which is not
empty in the final 5 elements results in elements being removed
from the view. So we can't remove the final 2 Ops. However,  in
```
  Chain c({1});
  c.settFillInto({10}, {{{{1, 9, 0}}}});
  Region r({10}, {{{{5,5,0}}}});
  c.settSample(r);
  c.settFillInto(r);
```
the final 2 Ops (settSample and settFillInto with `r`) can be removed, as
the the preceding settFillInto guarantees the sliced region of `r` contains
all the input elements.

This diff implements this simplification pass.

Test Plan: New tests added

Reviewers: shirazb, #poprithms

Reviewed By: shirazb, #poprithms

Maniphest Tasks: T52316

Differential Revision: https://phabricator.sourcevertex.net/D57691