[unwind] ExpandOp and unwinding through multiple SumLike inputs
Summary:
Expand is the only view-changing Op which is not surjective (1-1
correspondence between input elements and output elements).
Non-surjectivity makes it hard to choose how an input tensor's layout
should be chosen based on the output's layout. In this diff, we take the
lowest slice of the output.

This ambiguity in mapping from the output layout to the input layout also
breaks the connection between the poprithms abstraction and the (poplibs)
backend, as the poplar tensor layout of the output of expand will not
correspond the poprithms output (unless the poplar output tensor is made up
of repeats of the lowest slice).

The second change in this diff is to allow unwinding through multiple input
indices of a sum-like operation, like Add. This change introduces a new
discrepancy between the (poplibs) backend and poprithms, as in practise the
layout of an add cannot correspond to the layouts of all of the inputs.
Multi-index unwinding is necessary to ensure optimal layouts in certain
cases when poplar, or any backend where the the decision of which input
layout is used is opaque.

Test Plan: new tests added

Reviewers: shirazb, #poprithms

Reviewed By: shirazb, #poprithms

Maniphest Tasks: T53025

Differential Revision: https://phabricator.sourcevertex.net/D58234