[poprithms::unwind] Specify the order in which Paths should be unwound and Ops 'grown'
Summary:
The poprithms unwinding Solution specifies which tensors should be
unwound to which tensors, but it does not specify the order in which paths
should be unwound, or the order in which ops should be 'grown'. This diff
adds an additional layer of abstraction above the Solution class which
computes the order, and encapsulated it in a new class, ScheduledSolution. In
addition, the mechanics for unwinding are captured in a template class,
which can be used by different backend (poplar, and a mock testing class).

Test Plan: A mock class is created and used to test the template lowering

Reviewers: shirazb, #poprithms

Reviewed By: shirazb, #poprithms

Maniphest Tasks: T53086

Differential Revision: https://phabricator.sourcevertex.net/D58268