Program decomposition into constituent ipu/host
Summary:
With poplar, a user specifies which poplar::Programs can be dynamically
executed from the host. This is done by a poplar::Engine constructor
parameter.

This diff considers generalized sequences of ops, where ops can execute
either on host or ipu (see the CodeLocation enum). As with poplar, the user
specifies which of the generalized sequences can be executed dynamically.

The goal of this project is to determine which ops in a generalized
sequence correspond to common poplar sequences, and which poplar programs
must be made dynamically executable.

Reviewers: #poprithms, shirazb

Reviewed By: #poprithms, shirazb

Subscribers: shirazb

Maniphest Tasks: T53833

Differential Revision: https://phabricator.sourcevertex.net/D58602