Methods to compute max (min) path length to terminal node
Summary:
Given a node in a DAG, what is the longest (or shortest) path from it to a node
which has no outputs (a terminal node)? The solution is found with a simple
dynamic programming algorithm run on the nodes, in reverse order of a valid
schedule. Max (longest) and min (shortest) are very similar, as is the statistic
which counts the total number of paths to a terminal node.

Test Plan: new tests added

Reviewers: shirazb, #poprithms

Reviewed By: shirazb, #poprithms

Maniphest Tasks: T53835

Differential Revision: https://phabricator.sourcevertex.net/D58898