[unwinding] Fix mock class's implemention Helper::finalLayout
Summary:
Before this fix, the mock class was indexing into a map using an unwind-graph
tensor id instead of a mock class tensor id. This is fixed in this diff.

In addition, a few more assertions are sprinkled into the core logic to detect
errors earlier.

Test Plan: A new test is added, which failed before this patch.

Reviewers: george, #poprithms

Reviewed By: george, #poprithms

Subscribers: fabiant

Maniphest Tasks: T54215

Differential Revision: https://phabricator.sourcevertex.net/D59160