[chain] Ability to replace slice->reshape with reshape->slice
Summary:
Is it possible replace
    tensor.slice(low0, upp0).reshape(sh0)
with,
   tensor.reshape(sh1).slice(low1, upp1) ?

That is, instead of slicing then reshaping a tensor, can you reshape and
then slice the tensor and still end up with the same result?

This diff adds methods to determine when this is possible and describes
sh1, low1 and upp1 are in terms of sh0, low0, and upp0.

Test Plan: new tests added.

Reviewers: george, #poprithms

Reviewed By: george, #poprithms

Maniphest Tasks: T54574

Differential Revision: https://phabricator.sourcevertex.net/D59429