[multiout] Ability to remove inputs/outputs from ops.
Summary:
For example, if op has 3 outputs with tensor ids {op, 0}, {op, 1}, {op, 2}, and
we want to remove output #1, we must

1:
shift the output {op, 2} down 1 index to {op, 1} so as to keep the indices contiguous
(recall that unlike popart, output indices must always be contiguous and non-optional)

2:
signal to the consumers of {op, 2} that they are actually now consumers of {op, 1}.

3:
what about the old consumers of the {op, 1} which is deleted? If there are any,
a substitute must be provided for them.

Similar story for input removal.

Test Plan: New tests added.

Reviewers: #poprithms, george

Reviewed By: #poprithms, george

Maniphest Tasks: T55383

Differential Revision: https://phabricator.sourcevertex.net/D60231