[schedule::scc] Ability to add edge information in summary
Summary:
Before this diff: only node/op information can be included in the summary of the
strongly connected components (cycle-in-graph reports). With this diff, a user
can add edge information to summaries. This will allow the user to see, for
example, if an edge is a data dependency or a control dependency, or whatever other
type of edge they might have.

Test Plan: new test added.

Reviewers: fabiant, #poprithms

Reviewed By: fabiant, #poprithms

Maniphest Tasks: T59149

Differential Revision: https://phabricator.sourcevertex.net/D63815