[multiout] Method to get forward edge map of a subset of ops
Summary:
A forward edge map for a multiout::Graph is basically just a map with key:values
pairs defining data dependencies.

In this diff, a new method is added to obtain partial forward maps of
multiout::Graphs. It has algorithmic complexity which scales with the size
of the connected components of the ops required in the forward edge map, rather
than scaling with the total number of ops in the graph, as the existing method does.

Test Plan: new tests added

Reviewers: shirazb, #poprithms

Reviewed By: shirazb, #poprithms

Maniphest Tasks: T59146

Differential Revision: https://phabricator.sourcevertex.net/D63812