[common::schedulable] more ways to insert constraints
Summary:
Currently, control dependencies (topo cons) is a catch-all category for
non-data dependant constraints in a schedulable::Graph. For example, they are
used to ensure correct behaviour when inplace ops are involved, and to reduce
tensor liveness. This diff allows users of this graph class to sub-classify
control dependencies, and include them for scheduling in different ways and at
different stages.

This diff also adds the ability for control dependencies to be transferred
to nearest neighbors. This is useful for example for ops which do not do any
computation such as view-changing ops, for which control dependencies might
not make sense.

Test Plan: New tests added

Reviewers: george, #poprithms

Reviewed By: george, #poprithms

Maniphest Tasks: T59332

Differential Revision: https://phabricator.sourcevertex.net/D63997