[common::compute] Basic graph and op classes
Summary:
This new poprithms namespace will be for a more general compute graph,
where tensors have numerical types and the abstractions for ops with
callees are introduced. It will be used for more comprehensive testing of
poprithms components. It should not contain anything but extremely simple
algorithms.

Reviewers: shirazb, #poprithms

Reviewed By: shirazb, #poprithms

Maniphest Tasks: T61455

Differential Revision: https://phabricator.sourcevertex.net/D66148