[common::compute] Container for tensors in numerical simulation
Summary:
This class will be used in a simulator run of a poprithms::common::compute::Graph.
It stores for every tensor (host, ipu, remote) a HostTensor, with an appropriate
replication factor for ipu tensors. It also stores some additional state to simulate
for example the pointer address in a circular buffer (for copying between host and ipu).

Test Plan: Test of cloning added, the rest of the code is trivial.

Reviewers: shirazb, #poprithms, romand

Reviewed By: #poprithms, romand

Subscribers: romand, matthewha

Maniphest Tasks: T62643

Differential Revision: https://phabricator.sourcevertex.net/D67429