[schedule::shift] Improve inner loop performance
Summary:
On a small gnn model, there is a ~20% overall improvement:

Times after [seconds] : 4.931 4.939 4.910 4.927
Times before [seconds]  : 5.886 5.895 5.864 5.830

The change is to make the inner loop of the scheduling algorithm a pure integer
problem, moving the vector of doubles representing a tensor's weight out of the
loop. It is not immediately obvious why this improves performance but the compiler
seems to like this... and (imo) it makes the code clearer to human eyes too.

Test Plan: No logical change, CI.

Reviewers: george, #poprithms

Reviewed By: george, #poprithms

Subscribers: jeromem

Maniphest Tasks: T63294

Differential Revision: https://phabricator.sourcevertex.net/D68181