[common::compute] Define how output values depend on input values
Summary:
It is useful to know if the output value of an op depends on an input value
when doing some transformations/optimizations.

In general all output values depend on all inputs. Two exceptions are

The op which zeros its input
The op which overwrites the values of one input with the values of another input.

Reviewers: shirazb, #poprithms

Reviewed By: shirazb, #poprithms

Maniphest Tasks: T66875

Differential Revision: https://phabricator.sourcevertex.net/D72330