[program][pipeline] Graph pipelining algorithm
Summary:
Generic algorithm with minimalist interfaces for creating sub-graphs
(programs) and ops. Interfaces done with c++ polymorphism.

The pipelined model created is not strictly SSA. It requires some tensors to be
accessible in multiple graphs (a program for stashing, a program for computing, etc.).
This is done with the virtual method (to be implemented by user of project) to access
a tensor created in 1 sub-graph from another sub-graph.

Test Plan:
Testing is done by implementing the interfaces for the common::compute::Graph
class. Multiple numerical tests can be run rapidly using this approach. There
are also tests that check that the number of inter-device copies is as
expected, stash sizes are as expected, etc.

Reviewers: #poprithms, romand

Reviewed By: #poprithms, romand

Subscribers: romand

Maniphest Tasks: T66691

Differential Revision: https://phabricator.sourcevertex.net/D72083