Method to obtain a tensor view from a ptr-offset-strides representation
Summary:
An implementation of torch.as_strided for tensor classes with poplar-like APIs.

Background: Numpy (and torch) tensors are 'views' in the ptr-offset-strides format.

This utility lowers from the numpy representation to a poplar representation.

a = np.arange(42).reshape((6,7))

     [[ 0  1  2  3  4  5  6  ]
      [ 7  8  9  10 11 12 13 ]
      [ 14 15 16 17 18 19 20 ]
      [ 21 22 23 24 25 26 27 ]
      [ 28 29 30 31 32 33 34 ]
      [ 35 36 37 38 39 40 41 ]]

Using numpy slice and transpose,

b = a[3:7:2,6:2:-3].T

      [[ 27 41 ],
       [ 24 38 ]]

Using the utility in the diff, starting from the strides, offset, and output shape:

strides = (-3, 14)
offset  = 27
shape   = (2,2)

The output tensor can be obtained directly, even when the chain of numpy view changes and starting shape have been lost:

poprithms::...::asStrided(a.flatten(), strides, offset, shape) creates same as above:

      [[ 27 41 ],
       [ 24 38 ]]

To use it, a class with a subset of the poplar::Tensor's API must be
implemented, and used as template parameter.

Reviewers: davidw, shirazb, #poprithms

Reviewed By: davidw, #poprithms

Maniphest Tasks: T70071

Differential Revision: https://phabricator.sourcevertex.net/D73530